Rework the example to use the new AWS Client
The previous version (v4) of aws-iot-device-sdk-embedded-C is now
deprecated in favour of coreMQTT from Amazon FreeRTOS.

An new AWS client based on coreMQTT and written for Mbed OS has been
contributed by Nantis GmbH, and this commit reworks the example to
support the new AWS client.

Two demos are provided. The demo is selected based on the config
in mbed_app.json:
* MQTT (default): if aws-client.shadow is unset or false
* Device Shadow service: if aws-client.shadow is true